+++ man/examples/examples_do_RankedEnrichmentHeatmap.R
@@ -9,9 +9,9 @@
     sample <- readRDS(system.file("extdata/seurat_dataset_example.rds", package = "SCpubr"))
     
     # Genes have to be unique.
-    genes <- list("Gene set A" = rownames(sample)[1:5],
-                  "Gene set B" = rownames(sample)[6:10],
-                  "Gene set C" = rownames(sample)[11:15])
+    genes <- list("A" = rownames(sample)[1:5],
+                  "B" = rownames(sample)[6:10],
+                  "C" = rownames(sample)[11:15])
     
     
     # This will query, for the provided components, the enrichment of the gene